COMP: fixes gcc6 error see https://issues.slicer.org/view.php?id=4413
git-svn-id: http://svn.slicer.org/Slicer3/branches/Slicer4-EMSegment@17136 3bd1e089-480b-0410-8dfb-8563597acbee

+++ AMF/vtkImagePropagateDist.cxx
@@ -61,8 +61,7 @@
 
 
 #define dist2(a,b,c)  (sq[abs(a)]+sq[abs(b)]+sq[abs(c)])
-#define dist(a,b,c)  (distance[abs(a)][abs(b)][abs(c)])
-
+#define dist(a,b,c)  (distance[(int)abs(a)][(int)abs(b)][(int)abs(c)])
 
 
 //
